pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "diagrams>=0.24.4",
    "graphviz>=0.20.3",
    "matplotlib>=3.10.5",
    "networkx>=3.5",
    "strands-agents[anthropic]>=1.4.0",
    "strands-agents-tools>=0.2.3",
    "python-dotenv>=1.1.1",
]
